IOS-8356 Add `OnrampRedirecting` module
