+++ Tangem/Modules/Send/Services/SendBaseDataBuilder/SendBaseDataBuilder.swift
@@ -24,7 +24,7 @@ protocol OnrampBaseDataBuilder: SendGenericBaseDataBuilder {
     func makeDataForOnrampCountryBottomSheet() -> OnrampRepository
     func makeDataForOnrampCountrySelectorView() -> (preferenceRepository: OnrampRepository, dataRepository: OnrampDataRepository)
     func makeDataForOnrampProvidersPaymentMethodsView() -> (providersBuilder: OnrampProvidersBuilder, paymentMethodsBuilder: OnrampPaymentMethodsBuilder)
-    func makeDataForOnrampWebView() -> OnrampManager
+    func makeDataForOnrampRedirecting() -> (tokenItem: TokenItem, onrampManager: OnrampManager)
 }
 
 protocol SendGenericBaseDataBuilder {
