--- Tangem/Modules/Send/UI/OnrampWebView/OnrampWebViewCoordinatorView.swift
@@ -1,38 +1,0 @@
-//
-//  OnrampWebViewCoordinatorView.swift
-//  Tangem
-//
-//  Created by Sergey Balashov on 07.11.2024.
-//  Copyright © 2024 Tangem AG. All rights reserved.
-//
-
-import SwiftUI
-
-struct OnrampWebViewCoordinatorView: CoordinatorView {
-    @ObservedObject var coordinator: OnrampWebViewCoordinator
-
-    init(coordinator: OnrampWebViewCoordinator) {
-        self.coordinator = coordinator
-    }
-
-    var body: some View {
-        ZStack {
-            if let rootViewModel = coordinator.rootViewModel {
-                OnrampWebViewView(viewModel: rootViewModel)
-                    .navigationLinks(links)
-            }
-
-            sheets
-        }
-    }
-
-    @ViewBuilder
-    private var links: some View {
-        EmptyView()
-    }
-
-    @ViewBuilder
-    private var sheets: some View {
-        EmptyView()
-    }
-}
