+++ Tangem/Modules/Send/UI/OnrampWebView/OnrampWebViewRoutable.swift
@@ -2,12 +2,12 @@
 //  OnrampWebViewRoutable.swift
 //  Tangem
 //
-//  Created by Sergey Balashov on 07.11.2024.
+//  Created by Sergey Balashov on 08.11.2024.
 //  Copyright © 2024 Tangem AG. All rights reserved.
 //
 
 import Foundation
 
 protocol OnrampWebViewRoutable: AnyObject {
-    func openURL(url: URL)
+    func dismissOnrampWebView()
 }
