+++ Tangem/Modules/Send/UI/OnrampWebView/OnrampWebViewViewModel.swift
@@ -2,59 +2,55 @@
 //  OnrampWebViewViewModel.swift
 //  Tangem
 //
-//  Created by Sergey Balashov on 07.11.2024.
+//  Created by Sergey Balashov on 08.11.2024.
 //  Copyright © 2024 Tangem AG. All rights reserved.
 //
 
 import Combine
 import SwiftUI
-import TangemExpress
 
 final class OnrampWebViewViewModel: ObservableObject {
     // MARK: - ViewState
 
-    var providerImageURL: URL? { onrampProvider.provider.imageURL }
-    var providerName: String { onrampProvider.provider.name }
+    var title: String { "\(Localization.commonBuy) \(tokenItem.name)" }
 
-    @Published var alert: AlertBinder?
+    var urlActions: [String: (String) -> Void] {
+        [
+            successURL: { [weak self] _ in
+                self?.coordinator?.dismissOnrampWebView()
+            }
+        ]
+    }
+
+    var url: URL { widgetURL }
 
     // MARK: - Dependencies
 
-    private let onrampProvider: OnrampProvider
-    private let onrampManager: OnrampManager
+    private let tokenItem: TokenItem
+    private let widgetURL: URL
+    private let successURL: String
     private weak var coordinator: OnrampWebViewRoutable?
 
     init(
-        onrampProvider: OnrampProvider,
-        onrampManager: OnrampManager,
+        settings: Settings,
         coordinator: OnrampWebViewRoutable
     ) {
-        self.onrampProvider = onrampProvider
-        self.onrampManager = onrampManager
+        self.tokenItem = settings.tokenItem
+        self.widgetURL = settings.widgetURL
+        self.successURL = settings.successURL
+
         self.coordinator = coordinator
     }
 
-    func loadRedirectData() async {
-        do {
-//            let redirectSettings = OnrampRedirectSettings(successURL: "https://tangem.com/onramp/success", theme: "light", language: "en")
-//            let item = try onrampProvider.manager.makeOnrampQuotesRequestItem()
-//            let requestItem = OnrampRedirectDataRequestItem(quotesItem: item, redirectSettings: redirectSettings)
-            let data = try await onrampManager.loadRedirectData(provider: onrampProvider)
-
-            guard let url = URL(string: data.widgetUrl) else {
-                throw Error.wrongURL
-            }
-
-            coordinator?.openURL(url: url)
-        } catch {
-            // TODO: close view ?
-            alert = error.alertBinder
-        }
+    func close() {
+        coordinator?.dismissOnrampWebView()
     }
 }
 
 extension OnrampWebViewViewModel {
-    enum Error: LocalizedError {
-        case wrongURL
+    struct Settings {
+        let tokenItem: TokenItem
+        let widgetURL: URL
+        let successURL: String
     }
 }
